Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
pydantic-settings
python-dotenv
cryptography
fastapi-mail 
python-jose[cryptography] 
passlib[bcrypt]
PyJWT
pydantic[email]